+++ src/Symfony/Bridge/Twig/Tests/Translation/TwigExtractorTest.php
@@ -60,6 +60,9 @@ class TwigExtractorTest extends TestCase
             ['{% trans from "domain" %}new key{% endtrans %}', ['new key' => 'domain']],
             ['{% set foo = "new key" | trans %}', ['new key' => 'messages']],
             ['{{ 1 ? "new key" | trans : "another key" | trans }}', ['new key' => 'messages', 'another key' => 'messages']],
+            ['{{ t("new key") | trans() }}', ['new key' => 'messages']],
+            ['{{ t("new key", {}, "domain") | trans() }}', ['new key' => 'domain']],
+            ['{{ 1 ? t("new key") | trans : t("another key") | trans }}', ['new key' => 'messages', 'another key' => 'messages']],
 
             // make sure 'trans_default_domain' tag is supported
             ['{% trans_default_domain "domain" %}{{ "new key"|trans }}', ['new key' => 'domain']],